python_backend/requirements.txt
Flask
tensorflow
numpy
Pillow
flask_cors